+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+[Check the diff](https://github.com/elastic/apm-agent-python/compare/v5.1.0...master)
+
+### Bugfixes
+ * fixed an issue with empty responses from APM Server's config endpoint (#562, #563)
+
 ## v5.1.0 
 [Check the diff](https://github.com/elastic/apm-agent-python/compare/v5.0.0...v5.1.0)